% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddOnSearch.R
\name{AddOnSearch}
\alias{AddOnSearch}
\title{Search the optimal model region from a close-to-optimum model}
\usage{
AddOnSearch(merge.code, varia.list, model, transition.method, IncludeOrigin)
}
\arguments{
\item{merge.code}{the given close-to-optimum graycode, list(code1, code2)}

\item{varia.list}{a list of variables, c(.., .., ..)}

\item{model}{the full model}
}
\value{
the optimal partitions
}
\description{
Given a close-to-optimum model, this function continuously searches the neighbour models to find the global optimum, through a complete search
}
\note{
check "addon.code.trace" and "addon.bic.trace" in global environment
}
\examples{
code1 = c(1,2,2,3,3)
code2 = c(1,2,3,4,5,2,6,7,8)
code1 = c(1,1,1,2,3)
code2 = c(1,2,3,3,5,2,6,7,7)

addon.code.trace <<- NULL
addon.bic.trace <<- NULL
AddOnSearch.function(merge.code = list(code1,code2),
            varia.list = c("Kilometres", "Make"),
            model = mod2.4)
addon.code.trace
addon.bic.trace

addon.code.trace <<- NULL
addon.bic.trace <<- NULL
AddOnSearch.function(merge.code = list(c(1,2,2,2,3)),
                     varia.list = c("Kilometres"),
                     model = mod2.4)
addon.code.trace
addon.bic.trace

resss <- rxAddOnSearch.function(final.merge.code = list(as.vector(Extract.SA.Res( countyfreq.res[[2]], 0.6)$State[[2]])),
varia.list = c("COUNTYNAME"),
mod=fin.freq.mod.rx)
}
