% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_model_weight.R
\name{bic.model.weight}
\alias{bic.model.weight}
\title{Approximate model posterior probability using BIC.}
\usage{
bic.model.weight(BIC.vec)
}
\arguments{
\item{BIC.vec}{a vector of all BIC values used to calculate model weight}
}
\value{
the approximated posterior model probability, i.e. model weight used in BMA.
}
\description{
This function allows you to calculate/approximate model posterior probability given BIC values, which can then be used for model weights in BMA weighted average.
}
\examples{
BIC_BMAweight(c(10,11))
}
\seealso{
\code{\link[BMA]{bma.glm}}
}
