% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_data.R
\name{collapse.data}
\alias{collapse.data}
\title{Reorganise dataset}
\usage{
collapse.data(data, varia.list, graycode.list)
}
\arguments{
\item{data}{the dataset to be reorganised/collapsed}

\item{varia.list}{variables to be collapsed}

\item{graycode.list}{graycodes}
}
\value{
new reorganised/collapsed dataset
}
\description{
Given dataset, variables to be collapsed, and corresponding graycode, this function returns the collapsed dataset
}
