% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCollapsing.R
\name{combineCollapsing}
\alias{combineCollapsing}
\title{Combining the best collapsings from factors collapsed separately}
\usage{
combineCollapsing(all.collapsings, varia.list, model)
}
\arguments{
\item{all.collapsings}{a list of list of all collapsing results for each factor}

\item{varia.list}{a vector of all factors collapsed}

\item{model}{full model}
}
\value{
the best combinations
}
\description{
When collapsing factors individually for the reason of computational intensity, this functions searches across the best results from each individual factor and find the best combinations across all factors
}
\examples{
\donttest{
r1 <- FCBMA(model = mod1,
            varia.list = c("Make"),
            method = "SA",
            transition.method = "ChangeOne")
res1 <- Extract.SA.Res(r1, threshold.val = 0.95)$State

r2 <- FCBMA(model = mod1,
            varia.list = c("Zone"),
            method = "SA",
            transition.method = "ChangeOne")
res2 <- Extract.SA.Res(r2, threshold.val = 0.95)$State

r3 <- FCBMA(model = mod1,
            varia.list = c("Kilometres"),
            method = "SA",
            transition.method = "ChangeOne")
res3 <- Extract.SA.Res(r3, threshold.val = 0.95)$State
z <- CombineCollapsing(all.collapsings = list(unlist(Res1$Accept.States, recursive = F),
                                              unlist(Res2$Accept.States, recursive = F),
                                              unlist(Res3$Accept.States, recursive = F)),
                                varia.list=c("Make", "Zone", "Kilometres"),
                                model=mod1)
}

}
