% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cano_graycode.R
\name{convert.cano.graycode}
\alias{convert.cano.graycode}
\title{Convert a given graycode into its canonical format}
\usage{
convert.cano.graycode(code)
}
\arguments{
\item{code}{the input graycode which can be of any random  integer strings}
}
\value{
the canonical graycode
}
\description{
Convert a given graycode into its canonical format: beginning with 1, digits representing different groups are of increasing order
}
\examples{
a <- convert.cano.graycode(c(2, 1, 3, 4, 2)); print(a)
b <- sample(c(1:9), 10000, replace = T)
system.time({b <- convert.cano.graycode(b)})
print(b)
}
