% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_canonical_graycode.R
\name{convert_canonical_graycode}
\alias{convert_canonical_graycode}
\title{Convert a given graycode into its canonical format}
\usage{
convert_canonical_graycode(code)
}
\arguments{
\item{code}{An interger vector, i.e. the input graycode, which can be of any random integer vectors}
}
\value{
An integer vector of the canonical graycode
}
\description{
Convert a given graycode into its canonical format: beginning with 1, digits representing different groups are of increasing order
}
\examples{
convert_canonical_graycode(c(2, 1, 3, 4, 2))
a <- sample(c(1:9), 1000, replace = TRUE)
convert_canonical_graycode(a)

}
