% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_level_collapsing.R
\name{factor.level.collapsing}
\alias{factor.level.collapsing}
\title{Factor collapsing over one factor}
\usage{
factor.level.collapsing(variable, graycode)
}
\arguments{
\item{variable}{a given variable}

\item{graycode}{a given graycode}
}
\value{
a collapsed version of the input variable
}
\description{
For a given (standard) gray code, and a variable, this function returns the collapsed version of this factor based on this given gray code
}
\examples{
factor.level.collapsing(data$Make, c(1,2,3,4,5,6,7,8,8))
}
