% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_predict.R
\name{fc.predict}
\alias{fc.predict}
\alias{fc.predict.glm}
\alias{fc.predict.rxGlm}
\title{Prediction based on FC}
\usage{
fc.predict(data, model, varia.list, merge.list)

\method{fc.predict}{glm}(data, model, varia.list, merge.list)

\method{fc.predict}{rxGlm}(data, model, varia.list, merge.list)
}
\arguments{
\item{data}{new dataset, such as testing data}

\item{model}{model, such as training model including interactions}

\item{varia.list}{variables to be collapsed}

\item{merge.list}{graycodes}
}
\value{
prediction
}
\description{
Prediction based on the fitted model given the collpase/partition that selected by running the stochastic search
}
\examples{

m1 <- glm(Claims ~ Kilometres+Zone+Bonus+Make, offset = log(Insured),
          data = sweden, family = "poisson")
a <- fc.predict(data=sweden,
                model = m1,
                varia.list = c("Kilometres", "Zone", "Bonus", "Make"),
                merge.list = list(c(1,2,3,4,4),
                                  c(1,2,3,4,5,5,5),
                                  c(1,1,1,2,3,4,5),
                                  c(1,2,3,4,5,6,7,7,7)) )
b <- predict.glm(m1, type='response')
plot(cbind(a,b))
}
