% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCBMA.R
\name{FCBMA}
\alias{FCBMA}
\title{Factor collapsing (FC) with Bayesian model averaging (BMA) with lm, glm, rxGlm (package "RevoScaleR")}
\usage{
FCBMA(model, varia.list, method = "complete",
  transition.method = "ChangeOne", IncludeOrigin = FALSE,
  AddOnSearch = FALSE, group = NULL, apart = NULL,
  control.SA = list(init.Temp = 10000, stop.Temp = 1e-05, coolingRate =
  0.7, Mtrial = 20), control.GA = list(popnSize = 20, CrossOverRate =
  0.8, MutationRate = 0.1, elitism = 0.1, MaxGen = 200), verbose = TRUE)
}
\arguments{
\item{model}{The model used for factor collapsing.}

\item{varia.list}{Vector of variables to be collapsed.}

\item{method}{Search method to use, choose from \code{"complete"}, \code{"SA"}, \code{"GA"}.}

\item{transition.method}{Transition method to change one graycode to its neighbouring graycode, choose from \code{"ChangeOne"}, \code{"GroupSplit"}. The default is \code{"ChangeOne"}.}

\item{IncludeOrigin}{Logical; When choose neighbouring graycode, should the input graycode be included or not.}

\item{AddOnSearch}{Logical; used when \code{method = "SA"} or \code{"GA"}. If \code{TRUE}, after stochastic search a short greedy search will be implemented to ensure the final result is indeed global optimum. The default is \code{FALSE}.}

\item{group}{A (list of) vector indicating which elements must be grouped together, the default is \code{NULL}.}

\item{apart}{A (list of) vector indicating which elements must not be grouped together, the default is \code{NULL}.}

\item{control.SA}{A list of control parameters for SA.
\describe{
\item{init.Temp}{Initial temperature, default value is 10000.}
\item{stop.Temp}{Stoping temperature, default value is 1e-5.}
\item{coolingRate}{Cooling rate, default value is 0.7.}
\item{Mtrial}{Number of interations at each temperature level, default is 20.}
}}

\item{control.GA}{A list of control parameters for GA.
\describe{
\item{popnSize}{Population size, detaulf is 20.}
\item{CrossOverRate}{Crossover rate, default is 0.8.}
\item{MutationRate}{Mutation rate, default is 0.1.}
\item{elitism}{Elitism rate, default is 0.1.}
\item{MaxGen}{Maximum generation allowed, default is 200.}
}}

\item{verbose}{Logical argument controlling whether progress will be printed while the search runs. Default is \code{TRUE}.}
}
\value{
A list with the elements
\describe{
\item{best.bic}{The best value of BIC.}
\item{best.state}{The best partitions for all stated variables.}
\item{all.bic}{All the searched BIC values.}
\item{all.states}{All the searched (combinations of) partitions, in graycode.}
\item{all.partitions}{All the searched (combinations of) partitions}
\item{all.weights}{All the BMA weights of searched partitions.}
\item{Table}{A summary table of best few BICs, partitions, BMA weights.}
\item{Accept.BIC}{Accepted BIC values along SA search; only when \code{method="SA"}.}
\item{Accept.States}{Accepted (combinations of) partitions along SA search; only when \code{method="SA"}.}
\item{Best.BIC.trace}{Trace of best BIC values along GA search; only when \code{method="GA"}.}
\item{AddOn.BestBIC}{The best BIC value found through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.BestState}{The best (combination of) partition found through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.BIC.trace}{The best BIC values found in each iteration through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.State.trace}{The best (combination of) partitions found in each iteration through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
}
}
\description{
This function use FC-BMA method to refit linear or generalized linear models, method chosen from complete exhaustive search, Simulated Annealing (SA), Generic Algorithm (GA).
}
\note{
This version of stochastic search changes on the simulated annealing part. It changed the stochastic search method at each iteration, only one variable is picked and change to one of its neighbours the probability of being picked is based on the number of levels You also have the option of which method to select a neighbour of a graycode
}
\examples{
data("sweden")
m1 <- glm(Claims ~ Kilometres+Zone+Bonus+Make, offset = log(Insured),
          data = sweden, family = "poisson")
summary(m1)

# complete search
m2 <- FCBMA(model = m1,
            varia.list = c("Kilometres"),
            method = "complete",
            verbose = FALSE)

\donttest{
# SA search
m3 <- FCBMA(model = m1,
            varia.list = c("Kilometres","Make"),
            method = "SA",
            transition.method = "ChangeOne",
            AddOnSearch = TRUE,
            control.SA = list(init.Temp = 1000,
                              coolingRate = 0.6,
                              stop.Temp = 1e-5,
                              Mtrial = 20),
            verbose = TRUE)
m3$best.state
m3$best.bic

# GA search
m4 <- FCBMA(model=m1,
            varia.list = c("Kilometres", "Make"),
            method = "GA",
            transition.method = "ChangeOne",
            AddOnSearch= TRUE,
            control.GA = list(popnSize=20,
                              CrossOverRate=0.8,
                              MutationRate=0.1,
                              elitism=0.1,
                              MaxGen=100),
            verbose = TRUE)
m4$best.state
m4$best.bic
}

}
