% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiniIndex.R
\name{Gini}
\alias{Gini}
\title{Gini index calculation}
\usage{
Gini(a, p, expo = NULL, plot = FALSE)
}
\arguments{
\item{a}{actual value}

\item{p}{prediction}

\item{expo}{exposure}

\item{plot}{logica; whether to plot}
}
\value{
Gini index
}
\description{
Gini index calculation
}
\examples{
Gini(a = a,
     p = p,
     expo = expo)
Gini(a = a,
     p = pred1,
     expo = expo)
a <- c(4,2,6,5,8)
p <- c(5,5,5,4,6)
Gini(a,p)/Gini(a,a)
}
