% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fc.R
\name{predict.fc}
\alias{predict.fc}
\alias{predict.fc.glm}
\alias{predict.fc.rxGlm}
\title{Prediction based on FC}
\usage{
predict.fc(data, model, varia.list, merge.list)

\method{predict}{fc.glm}(data, model, varia.list, merge.list)

\method{predict}{fc.rxGlm}(data, model, varia.list, merge.list)
}
\arguments{
\item{data}{new dataset, such as testing data}

\item{model}{model, such as training model including interactions}

\item{varia.list}{variables to be collapsed}

\item{merge.list}{graycodes}
}
\value{
prediction
}
\description{
Prediction based on the fitted model given the collpase/partition that selected by running the stochastic search
}
\section{Methods (by class)}{
\itemize{
\item \code{fc.glm}: for standard GLM

\item \code{fc.rxGlm}: for class rxGlm in RevoScaleR package
}}

\examples{
a <-predict.fc(data=dataset,
               model = mod1.1,
               varia.list = c("Kilometres", "Zone", "Bonus", "Make"),
               merge.list = list(c(1,2,3,4,4),
                                 c(1,2,3,4,5,5,5),
                                 c(1,1,1,2,3,4,5),
                                 c(1,2,3,4,5,6,7,7,7)) )
b <- predict.glm(new.mod, type='response')
}
