% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxAICtoBIC.R
\name{rxAICtoBIC}
\alias{rxAICtoBIC}
\title{convert from AIC to BIC, from a rxGlm model}
\usage{
rxAICtoBIC(rxmodel, rxdata)
}
\arguments{
\item{rxmodel}{a rxGlm model}

\item{rxdata}{has to be the data associated with rxmode}
}
\value{
BIC and loglikelihood
}
\description{
convert from AIC to BIC, from a rxGlm model
}
