% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPartitions.R
\name{setPartitions}
\alias{setPartitions}
\alias{setPartitions.restricted}
\title{Generate set partitions}
\usage{
setPartitions(num)

setPartitions.restricted(num, group = NULL, apart = NULL)
}
\arguments{
\item{num}{the number of elements in a set}

\item{group}{a vector indicating which elements should be grouped together}

\item{apart}{a vector indicating which elements should not be grouped together}
}
\value{
a matrix of all set partitions
}
\description{
Enumeration of set partitions given the number of elements in a set
}
\note{
This function is based on the \code{\link[partitions]{setparts}} function from \code{"partitions"} package (R. K. S. Hankin) with modifications for FCBMA problem.
}
\examples{
setPartitions(4)
setPartitions.restricted(num=4, group=NULL, apart=c(2,3))
}
\references{
R. K. S. Hankin 2007. "Set partitions in R". Journal of Statistical Software, Volume 23, code snippet 2

R. K. S. Hankin 2006. Additive integer partitions in R. Journal of Statistical Software, Code Snippets 16(1)

Hu, S., O'Hagan, A. and Murphy, T. B. (2007). Motor Insurance Accidental Damage Claims Modelling with Factor Collapsing and Bayesian model Averaging
}
