% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitAt.R
\name{SplitAt}
\alias{SplitAt}
\title{Split a vector at a certain location/position}
\usage{
SplitAt(x, pos)
}
\arguments{
\item{x}{a vector}

\item{pos}{position within the vector x}
}
\value{
a list of splited vectors
}
\description{
Split a vector at a certain location/position
}
\examples{
x <- c(1,2,3,4,5)
SplitAt(x, 3)

}
