% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FCBMA.R
\name{summary.FCBMA}
\alias{summary.FCBMA}
\alias{print.summary.FCBMA}
\title{Summarizing FCBMA method}
\usage{
\method{summary}{FCBMA}(object, ...)

\method{print}{summary.FCBMA}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class "\code{FCBMA}" resulting of a call to \code{FCBMA}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "\code{summary.FCBMA}", usually a result of a call to \code{summary.FCBMA}.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
Summary method for class "\code{FCBMA}"
}
\examples{
\donttest{
data("sweden")
m1 <- glm(Claims ~ Kilometres+Zone+Bonus+Make, offset = log(Insured),
          data = sweden, family = "poisson")
summary(m1)

m2 <- FCBMA(m1,
            varia.list = c("Kilometres"),
            method = "complete",
            verbose = FALSE)
summary(m2)
}

}
