% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TournamentSelection.R
\name{TournamentSelection}
\alias{TournamentSelection}
\title{Tournament Selection}
\usage{
TournamentSelection(num, k, fitvec)
}
\arguments{
\item{num}{number of values required}

\item{k}{number of selection each time}

\item{fitvec}{vector of fitness value of all elements to be selected}
}
\value{
positions of elements selected
}
\description{
Tournament Selection
}
\note{
This tournament selection is written for function minimisation
}
\examples{
fitvec <- rnorm(10000, 0, 100)
TournamentSelection(200, 5, fitvec)
}
