% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddOnSearch.R
\name{AddOnSearch}
\alias{AddOnSearch}
\alias{AddOnSearch.glm}
\alias{AddOnSearch.rxGlm}
\title{Search the optimal model region from a close-to-optimum model}
\usage{
AddOnSearch(merge.code, varia.list, model, transition.method,
  IncludeOrigin = FALSE)

\method{AddOnSearch}{glm}(merge.code, varia.list, model, transition.method,
  IncludeOrigin = FALSE)

\method{AddOnSearch}{rxGlm}(merge.code, varia.list, model,
  transition.method, IncludeOrigin = FALSE)
}
\arguments{
\item{merge.code}{the given close-to-optimum graycode, list(code1, code2)}

\item{varia.list}{a list of variables, c(.., .., ..)}

\item{model}{the full model}

\item{transition.method}{Transition method to change one graycode to its neighbouring graycode, choose from \code{"ChangeOne"}, \code{"GroupSplit"}. The default is \code{"ChangeOne"}.}

\item{IncludeOrigin}{Logical; When choose neighbouring graycode, should the input graycode be included or not.}
}
\value{
the optimal partitions
}
\description{
Given a close-to-optimum model, this function continuously searches the neighbour models to find the global optimum, through a complete search
}
\examples{
\donttest{
data("sweden")
m1 <- glm(Claims ~ Kilometres+Zone+Bonus+Make, offset = log(Insured),
          data = sweden, family = "poisson")

code1 <- c(1,2,2,3,3)
code2 <- c(1,2,3,4,5,2,6,7,8)
assign("addon.bic.trace", NULL, envir = environment(FCBMA))
assign("addon.code.trace", NULL, envir = environment(FCBMA))
AddOnSearch(merge.code = list(code1,code2),
            varia.list = c("Kilometres", "Make"),
            model = m1,
            transition.method = "ChangeOne")
get("addon.bic.trace", envir = environment(FCBMA))
get("addon.code.trace", envir = environment(FCBMA))

assign("addon.bic.trace", NULL, envir = environment(FCBMA))
assign("addon.code.trace", NULL, envir = environment(FCBMA))
AddOnSearch(merge.code = list(c(1,2,2,2,3)),
            varia.list = c("Kilometres"),
            model = m1,
            transition.method = "ChangeOne")
get("addon.bic.trace", envir = environment(FCBMA))
get("addon.code.trace", envir = environment(FCBMA))
}

}
