% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_model_refit.R
\name{fc.model.refit}
\alias{fc.model.refit}
\alias{fc.model.refit.glm}
\alias{fc.model.refit.rxGlm}
\alias{fc.model.refit.lm}
\title{Refit a model based on collapsed factors}
\usage{
fc.model.refit(varia.list, merge.list, mod)

\method{fc.model.refit}{glm}(varia.list, merge.list, mod)

\method{fc.model.refit}{rxGlm}(varia.list, merge.list, mod)

\method{fc.model.refit}{lm}(varia.list, merge.list, mod)
}
\arguments{
\item{varia.list}{a vector of names of factors to be collapsed}

\item{merge.list}{a list of graycodes, each corresponding to factor in the varia.list}

\item{mod}{the model to be refitted}
}
\value{
A list with the elements
\item{BIC}{The BIC value of the refitted model}
\item{AIC}{The AIC value of the refitted model}
\item{logLik}{The log-likelihood of the refitted model}
\item{model}{Details of the refitted model}
}
\description{
Refit a model based on collapsed factors
}
\examples{
data("sweden")
m1 <- glm(Claims ~ Kilometres+Zone+Bonus+Make, offset = log(Insured),
          data = sweden, family = "poisson")
m2 <- fc.model.refit(varia.list=c("Kilometres"),
                     merge.list=list(c(1,2,3,3,4)),
                     mod = m1)
m3 <- fc.model.refit(varia.list=c("Kilometres, Make"),
                     merge.list=list(c(1,2,3,4,5), c(1,2,3,4,5,6,7,8,8)),
                     mod = m1)
}
