% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCBMA.R
\name{FCBMA}
\alias{FCBMA}
\title{Factor collapsing (FC) with Bayesian model averaging (BMA) with lm, glm, rxGlm (package "RevoScaleR")}
\usage{
FCBMA(model, varia.list, method = c("complete", "SA", "GA"),
  transition.method = "ChangeOne", IncludeOrigin = FALSE,
  AddOnSearch = FALSE, group = NULL, apart = NULL, cut = FALSE,
  cutoff = 0.8, Temp = 10000, coolingRate = 0.7, stop.Temp = 1e-05,
  Mtrial = 20, popnSize = 20, CrossOverRate = 0.8, MutationRate = 0.1,
  elitism = 0.1, MaxGen = 200, verbose = TRUE)
}
\arguments{
\item{model}{The model used for factor collapsing.}

\item{varia.list}{Vector of variables to be collapsed.}

\item{method}{Search method to use, choose from \code{"complete"}, \code{"SA"}, \code{"GA"}.}

\item{transition.method}{Transition method to change one graycode to its neighbouring graycode, choose from \code{"ChangeOne"}, \code{"GroupSplit"}. The default is \code{"ChangeOne"}.}

\item{IncludeOrigin}{Logical; When choose neighbouring graycode, should the input graycode be included or not.}

\item{AddOnSearch}{Logical; used when \code{method = "SA"} or \code{"GA"}. If \code{TRUE}, after stochastic search a short greedy search will be implemented to ensure the final result is indeed global optimum. The default is \code{FALSE}.}

\item{group}{A (list of) vector indicating which elements must be grouped together, the default is \code{NULL}.}

\item{apart}{A (list of) vector indicating which elements must not be grouped together, the default is \code{NULL}.}

\item{cut}{Logical; if \code{FALSE}, all searched results will be presented. If \code{TRUE}, only collapsed results above a certain cumulative BMA weights threshold will be presented. The default is \code{FALSE}.}

\item{cutoff}{Only when \code{cut=TRUE}; threshold value of accummulated best few model weights, the default is 0.8}

\item{Temp}{SA parameter, initial temperature, default value is 10000.}

\item{coolingRate}{SA parameter, cooling rate, default value is 0.7.}

\item{stop.Temp}{SA parameter, stoping temperature, default value is 1e-5.}

\item{Mtrial}{SA parameter, number of interations at each temperature level, default is 20.}

\item{popnSize}{GA parameter, population size, detaulf is 20.}

\item{CrossOverRate}{GA parameter, crossover rate, default is 0.8.}

\item{MutationRate}{GA parameter, mutation rate, default is 0.1.}

\item{elitism}{GA parameter, elitism rate, default is 0.1.}

\item{MaxGen}{GA parameter, maximum generation allowed, default is 200.}

\item{verbose}{Logical argument controlling whether progress will be printed while the search runs. Default is \code{TRUE}.}
}
\value{
A list with the elements
\describe{
\item{Best.Bic}{The best value of BIC.}
\item{Best.State}{The best partitions for all stated variables.}
\item{All.BIC}{All the searched BIC values.}
\item{All.States}{All the searched (combinations of) partitions.}
\item{Table}{A summary table of best few BICs, partitions, BMA weights.}
\item{Accept.BIC}{Accepted BIC values along SA search; only when \code{method="SA"}.}
\item{Accept.States}{Accepted (combinations of) partitions along SA search; only when \code{method="SA"}.}
\item{Best.BIC.trace}{Trace of best BIC values along GA search; only when \code{method="GA"}.}
\item{AddOn.BestBIC}{The best BIC value found through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.BestState}{The best (combination of) partition found through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.BIC.trace}{The best BIC values found in each iteration through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
\item{AddOn.State.trace}{The best (combination of) partitions found in each iteration through the add-on greedy search, if \code{AddOnSearch=TRUE}.}
}
}
\description{
This function use FC-BMA method to refit linear or generalized linear models, method chosen from complete exhaustive search, Simulated Annealing (SA), Generic Algorithm (GA).
}
\note{
This version of stochastic search changes on the simulated annealing part. It changed the stochastic search method at each iteration, only one variable is picked and change to one of its neighbours the probability of being picked is based on the number of levels You also have the option of which method to select a neighbour of a graycode
}
\examples{
# get the Sweden TP insurance data set
library(faraway)
data("motorins")
dataset <- motorins
dataset$Bonus <- as.factor(dataset$Bonus)
dataset$Kilometres <- factor(dataset$Kilometres, ordered=F)

# Frequency Poisson GLM
freq1 <- glm(Claims~Zone+Bonus+Make+Kilometres,
             offset=log(Insured),
             data=dataset,
             family="poisson")
summary(freq1)

freq.comp <- FCBMA(freq1,
                    varia.list = c("Kilometres"),
                    method = "complete",
                    verbose = TRUE)
system.time({
  freq.sa <- FCBMA(model=freq1,
                   varia.list = c("Kilometres","Make"),
                   method = "SA",
                   transition.method = "ChangeOne",
                   AddOnSearch = TRUE,
                   Temp = 1000,
                   coolingRate = 0.6,
                   stop.Temp = 1e-5,
                   Mtrial = 20,
                   verbose = TRUE)
})
freq.sa$Best.State
freq.sa$Best.BIC
system.time({
  freq.ga <- FCBMA(model=freq1,
                   varia.list = c("Make"),
                   method = "GA",
                   transition.method = "ChangeOne",
                   popnSize=20,CrossOverRate=0.8,MutationRate=0.1,
                   elitism=0.1,
                   MaxGen=100,
                   verbose = TRUE)
})
freq.ga$Best.State
freq.ga$Best.BIC

}
